feat: Adding components to their own files
