+++ src/stories/Components/ProposalCard/ProposalCard.tsx
@@ -1,20 +1,15 @@
 import styled from 'styled-components';
-import { useParams } from 'react-router';
-import { isDesktop } from 'react-device-detect';
-import { FiArrowRight, FiCircle } from 'react-icons/fi';
 
-import { Box } from '../common/Layout';
-import ProposalStatus from '../ProposalStatus';
-import { Heading } from '../common/Typography';
+import { Box } from '../../../components/Guilds/common/Layout';
+import ProposalStatus from '../../../components/Guilds/ProposalStatus';
+import { Heading } from '../../../components/Guilds/common/Typography';
 import 'react-loading-skeleton/dist/skeleton.css';
-import UnstyledLink from '../common/UnstyledLink';
-import { useProposal } from 'hooks/Guilds/ether-swr/guild/useProposal';
-import useENSAvatar from '../../../hooks/Guilds/ether-swr/ens/useENSAvatar';
-import Avatar from '../Avatar';
-import { MAINNET_ID } from '../../../utils/constants';
+import UnstyledLink from '../../../components/Guilds/common/UnstyledLink';
+import Avatar from '../../../components/Guilds/Avatar';
 import { shortenAddress } from '../../../utils';
-import { Loading } from '../common/Loading';
-import useVoteSummary from 'hooks/Guilds/useVoteSummary';
+import { Loading } from '../../../components/Guilds/common/Loading';
+import ProposalCardVotes from './ProposalCardVotes';
+import ProposalCardActionSummary from './ProposalCardActionSummary';
 
 const CardWrapper = styled(Box)`
   border: 1px solid ${({ theme }) => theme.colors.muted};
@@ -66,48 +61,13 @@ const Detail = styled(Box)`
   margin-left: 0.5rem;
 `;
 
-const Icon = styled.img<{
-  spaceLeft?: boolean;
-  spaceRight?: boolean;
-  bordered: boolean;
-}>`
-  width: 1.5rem;
-  height: 1.5rem;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  ${props => props.spaceLeft && `margin-left: 0.5rem;`}
-  ${props => props.spaceRight && `margin-right: 0.5rem;`}
-
-  ${props =>
-    props.bordered &&
-    `
-    border: 1px solid #000;
-    border-radius: 50%;
-  `}
-`;
-
-const BorderedIconDetailWrapper = styled(IconDetailWrapper)`
-  border: 1px solid ${({ theme }) => theme.colors.border.initial};
-  border-radius: 1rem;
-  padding: 0.25rem 0.8rem;
-  flex: none;
-  display: flex;
-`;
-
 const ProposalStatusWrapper = styled.div`
   display: flex;
   flex: 1;
   justify-content: flex-end;
 `;
 
-interface ProposalCardProps {
-  id?: string;
-  href?: string;
-}
-
-const ProposalCard: React.FC<ProposalCardProps> = ({ id, href }) => {
+const ProposalCard: React.FC = () => {
   return (
     <UnstyledLink to={href || '#'}>
       <CardWrapper>
